fix minor

+++ src/FileIO/utilities.jl
@@ -47,7 +47,7 @@ end
 """
 file tfw
 """
-function savetfw(output::String, GSD, lx::Float64, uy::Float64)
+function savetfw(output::String, GSD::Float64, lx::Float64, uy::Float64)
 	fname = split(output,".")[1]
 	io = open(fname*".tfw","w")
 	write(io, "$(Float64(GSD))\n")
